Move the marks handling logic inside YearRangeSlier component
